backend_app/requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
discord.py==2.3.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
Pillow==10.1.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==2.1.0
yarl==1.9.3
